refactor: move each projection raycast into its transform file
