+++ src/render/terrain_raycast.bench.ts
@@ -3,11 +3,10 @@ import Point from '@mapbox/point-geometry';
 import {DEMData} from '../data/dem_data.ts';
 import {LngLat} from '../geo/lng_lat.ts';
 import {MercatorCoordinate} from '../geo/mercator_coordinate.ts';
-import {MercatorTransform} from '../geo/projection/mercator_transform.ts';
+import {MercatorTransform, raycastTerrainMercator} from '../geo/projection/mercator_transform.ts';
 import {OverscaledTileID} from '../tile/tile_id.ts';
 import {RGBAImage} from '../util/image.ts';
 import {Terrain} from './terrain.ts';
-import {raycastTerrainMercator} from './terrain_raycast.ts';
 import type {Tile} from '../tile/tile.ts';
 import type {Painter} from './painter.ts';
 import type {TileManager} from '../tile/tile_manager.ts';
