+++ src/render/terrain_raycast.test.ts
@@ -5,11 +5,11 @@ import {EXTENT} from '../data/extent.ts';
 import {LngLat} from '../geo/lng_lat.ts';
 import {MercatorCoordinate} from '../geo/mercator_coordinate.ts';
 import {GlobeTransform} from '../geo/projection/globe_transform.ts';
-import {MercatorTransform} from '../geo/projection/mercator_transform.ts';
+import {MercatorTransform, raycastTerrainMercator} from '../geo/projection/mercator_transform.ts';
+import {raycastTerrainGlobe} from '../geo/projection/vertical_perspective_transform.ts';
 import {OverscaledTileID} from '../tile/tile_id.ts';
 import {RGBAImage} from '../util/image.ts';
 import {Terrain} from './terrain.ts';
-import {raycastTerrainGlobe, raycastTerrainMercator} from './terrain_raycast.ts';
 import type {Tile} from '../tile/tile.ts';
 import type {Painter} from './painter.ts';
 import type {TileManager} from '../tile/tile_manager.ts';
